Fix filter types
